requirements.txt
lxml
python-dotenv